requirements.txt
Django>=5.2.8
djangorestframework>=3.15.2
supabase>=2.10.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
